+++ app/src/main/java/com/example/jokbalmanager/model/db/OrderDao.kt
@@ -17,11 +17,11 @@ interface OrderDao {
     fun findOrderByType(date: String, type: Int): OrderEntity?
 
     @Query("UPDATE orders SET weight = weight + :weight, deposit = deposit + :deposit WHERE date = :date AND type = :type")
-    fun addSameDate(date: String, type: Int, weight: Double, deposit: Int)
+    fun addSameDate(date: String, type: Int, weight: Double, deposit: Long)
 
     @Delete
     fun deleteOrder(order: OrderEntity)
 
     @Query("UPDATE orders SET date = :date, price = :price, weight = :weight, deposit = :deposit  WHERE date = :prevDate AND type = :type")
-    fun updateOrder(prevDate: String, type: Int, date: String, price: Int, weight: Double, deposit: Int)
+    fun updateOrder(prevDate: String, type: Int, date: String, price: Long, weight: Double, deposit: Long)
 }
